requirements.txt
streamlit==1.42.0
youtube-transcript-api==1.2.4
yt-dlp==2024.12.23
easyocr==1.7.2
opencv-python-headless==4.10.0.84
Pillow==11.1.0
requests==2.32.3
numpy==1.26.4
playwright==1.49.1
